@@ -35,15 +35,15 @@ class QuillCheckboxPointState extends State<QuillCheckboxPoint> {
     final fillColor = widget.value
         ? (widget.enabled
             ? theme.colorScheme.primary
-            : theme.colorScheme.onSurface.withOpacity(0.5))
+            : theme.colorScheme.onSurface.withValues(alpha: 0.5))
         : theme.colorScheme.surface;
     final borderColor = widget.value
         ? (widget.enabled
             ? theme.colorScheme.primary
-            : theme.colorScheme.onSurface.withOpacity(0))
+            : theme.colorScheme.onSurface.withValues(alpha: 0))
         : (widget.enabled
-            ? theme.colorScheme.onSurface.withOpacity(0.5)
-            : theme.colorScheme.onSurface.withOpacity(0.3));
+            ? theme.colorScheme.onSurface.withValues(alpha: 0.5)
+            : theme.colorScheme.onSurface.withValues(alpha: 0.3));
     final child = Container(
       alignment: AlignmentDirectional.centerEnd,
       padding: EdgeInsetsDirectional.only(end: widget.size / 2),
