@@ -260,7 +260,7 @@ class DefaultStyles {
 
     final inlineCodeStyle = TextStyle(
       fontSize: 14,
-      color: themeData.colorScheme.primary.withOpacity(0.8),
+      color: themeData.colorScheme.primary.withValues(alpha: 0.8),
       fontFamily: fontFamily,
     );
 
@@ -424,7 +424,7 @@ class DefaultStyles {
           defaultTextStyle.style.copyWith(
             fontSize: 20,
             height: 1.5,
-            color: Colors.grey.withOpacity(0.6),
+            color: Colors.grey.withValues(alpha: 0.6),
           ),
           baseHorizontalSpacing,
           VerticalSpacing.zero,
@@ -439,7 +439,7 @@ class DefaultStyles {
         null,
       ),
       quote: DefaultTextBlockStyle(
-        TextStyle(color: baseStyle.color!.withOpacity(0.6)),
+        TextStyle(color: baseStyle.color!.withValues(alpha: 0.6)),
         baseHorizontalSpacing,
         baseVerticalSpacing,
         const VerticalSpacing(6, 2),
@@ -451,7 +451,7 @@ class DefaultStyles {
       ),
       code: DefaultTextBlockStyle(
           TextStyle(
-            color: Colors.blue.shade900.withOpacity(0.9),
+            color: Colors.blue.shade900.withValues(alpha: 0.9),
             fontFamily: fontFamily,
             fontSize: 13,
             height: 1.15,
