@@ -1,17 +1,17 @@
-import 'dart:async' show Completer;
-import 'dart:ui' as ui;
-
 import 'package:flutter/cupertino.dart' show showCupertinoModalPopup;
-import 'package:flutter/foundation.dart' show kIsWeb, Uint8List;
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_quill/flutter_quill.dart'
     show ImageUrl, QuillController, StyleAttribute, getEmbedNode;
 import 'package:flutter_quill/internal.dart';
+import 'package:path/path.dart' as p;
+import 'package:url_launcher/url_launcher.dart';
 
 import '../../common/utils/element_utils/element_utils.dart';
 import '../../common/utils/string.dart';
-import '../../common/utils/utils.dart';
 import 'config/image_config.dart';
+import 'image_load_utils.dart';
+import 'image_save_utils.dart';
 import 'widgets/image.dart' show ImageTapWrapper, getImageStyleString;
 import 'widgets/image_resizer.dart' show ImageResizer;
 
@@ -23,6 +23,7 @@ class ImageOptionsMenu extends StatelessWidget {
     required this.imageSize,
     required this.readOnly,
     required this.imageProvider,
+    this.prefersGallerySave = true,
     super.key,
   });
 
@@ -32,6 +33,19 @@ class ImageOptionsMenu extends StatelessWidget {
   final ElementSize imageSize;
   final bool readOnly;
   final ImageProvider imageProvider;
+
+  // TODO(quill_native_bridge): Update this doc comment once saveImageToGallery()
+  //  is supported on Windows too (will be applicable like macOS). See https://pub.dev/packages/quill_native_bridge#-features
+  /// Determines if the image should be saved to the gallery instead of using the
+  /// system file save dialog for platforms that support both.
+  ///
+  /// Currently, the only platform where this applies is macOS.
+  ///
+  /// This is silently ignored on platforms that only support gallery save (Android and iOS)
+  /// or only image save.
+  ///
+  /// For more details, refer to [quill_native_bridge Saving images](https://pub.dev/packages/quill_native_bridge#-saving-images).
+  final bool prefersGallerySave;
 
   @override
   Widget build(BuildContext context) {
@@ -90,7 +104,9 @@ class ImageOptionsMenu extends StatelessWidget {
                 getImageStyleString(controller),
               );
 
-              final imageBytes = await _loadImageBytesFromImageProvider();
+              final imageBytes = await ImageLoader.instance
+                  .loadImageBytesFromImageProvider(
+                      imageProvider: imageProvider);
               if (imageBytes != null) {
                 await ClipboardServiceProvider.instance.copyImage(imageBytes);
               }
@@ -126,48 +142,90 @@ class ImageOptionsMenu extends StatelessWidget {
                 await config.onImageRemovedCallback.call(imageSource);
               },
             ),
-          if (!kIsWeb)
-            ListTile(
-              leading: const Icon(Icons.save),
-              title: Text(context.loc.save),
-              onTap: () async {
-                final messenger = ScaffoldMessenger.of(context);
-                final localizations = context.loc;
-                Navigator.of(context).pop();
+          ListTile(
+            leading: const Icon(Icons.save),
+            title: Text(context.loc.save),
+            onTap: () async {
+              final messenger = ScaffoldMessenger.of(context);
+              final localizations = context.loc;
+              Navigator.of(context).pop();
 
-                final saveImageResult = await saveImage(
+              SaveImageResult? result;
+              try {
+                result = await ImageSaver.instance.saveImage(
                   imageUrl: imageSource,
+                  imageProvider: imageProvider,
+                  prefersGallerySave: prefersGallerySave,
                 );
-                final imageSavedSuccessfully = saveImageResult.error == null;
+              } on GalleryImageSaveAccessDeniedException {
+                messenger.showSnackBar(SnackBar(
+                    content: Text(
+                  localizations.saveImagePermissionDenied,
+                )));
+                return;
+              }
 
-                messenger.clearSnackBars();
+              if (result == null) {
+                messenger.showSnackBar(SnackBar(
+                    content: Text(
+                  localizations.errorUnexpectedSavingImage,
+                )));
+                return;
+              }
 
-                if (!imageSavedSuccessfully) {
-                  messenger.showSnackBar(SnackBar(
-                      content: Text(
-                    localizations.errorWhileSavingImage,
-                  )));
+              if (kIsWeb) {
+                messenger.showSnackBar(SnackBar(
+                    content: Text(localizations.successImageDownloaded)));
+                return;
+              }
+
+              if (result.isGallerySave) {
+                messenger.showSnackBar(SnackBar(
+                  content: Text(localizations.successImageSavedGallery),
+                  action: SnackBarAction(
+                    label: localizations.openGallery,
+                    onPressed: () =>
+                        QuillNativeProvider.instance.openGalleryApp(),
+                  ),
+                ));
+                return;
+              }
+
+              if (isDesktopApp) {
+                final imageFilePath = result.imageFilePath;
+                if (imageFilePath == null) {
+                  // User canceled the system save dialog.
                   return;
-                }
-
-                var message = switch (saveImageResult.method) {
-                  SaveImageResultMethod.network =>
-                    localizations.savedUsingTheNetwork,
-                  SaveImageResultMethod.localStorage =>
-                    localizations.savedUsingLocalStorage,
-                };
-
-                if (isDesktopApp) {
-                  message = localizations.theImageHasBeenSavedAt(imageSource);
                 }
 
                 messenger.showSnackBar(
                   SnackBar(
-                    content: Text(message),
+                    content: Text(localizations.successImageSaved),
+                    // On macOS the app only has access to the picked file from the system save
+                    // dialog and not the directory where it was saved.
+                    // Opening the directory of that file requires entitlements on macOS
+                    // See https://pub.dev/packages/url_launcher#macos-file-access-configuration
+                    // Open the saved image file instead of the directory
+                    action: defaultTargetPlatform == TargetPlatform.macOS
+                        ? SnackBarAction(
+                            label: localizations.openFile,
+                            onPressed: () => launchUrl(Uri.file(imageFilePath)),
+                          )
+                        : SnackBarAction(
+                            label: localizations.openFileLocation,
+                            onPressed: () => launchUrl(
+                                Uri.directory(p.dirname(imageFilePath))),
+                          ),
                   ),
                 );
-              },
-            ),
+
+                return;
+              }
+
+              throw StateError(
+                  'Image save result is not handled on $defaultTargetPlatform');
+            },
+          ),
           ListTile(
             leading: const Icon(Icons.zoom_in),
             title: Text(context.loc.zoom),
@@ -184,24 +242,5 @@ class ImageOptionsMenu extends StatelessWidget {
         ],
       ),
     );
-  }
-
-  // TODO: This will load the image again, in case it was network image
-  //  then it will send a GET request each time to load the image.
-  Future<Uint8List?> _loadImageBytesFromImageProvider() async {
-    final stream = imageProvider.resolve(ImageConfiguration.empty);
-    final completer = Completer<ui.Image>();
-
-    ImageStreamListener? listener;
-    listener = ImageStreamListener((info, _) {
-      completer.complete(info.image);
-      stream.removeListener(listener!);
-    });
-
-    stream.addListener(listener);
-
-    final image = await completer.future;
-    final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
-    return byteData?.buffer.asUint8List();
   }
 }
