@@ -6,7 +6,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:url_launcher/url_launcher_string.dart' show launchUrlString;
+import 'package:url_launcher/url_launcher.dart';
 
 import '../../../../flutter_quill.dart';
 import '../../../common/utils/color.dart';
@@ -612,7 +612,7 @@ class _TextLineState extends State<TextLine> {
   }
 
   Future<void> _launchUrl(String url) async {
-    await launchUrlString(url);
+    await launchUrl(Uri.parse(url));
   }
 
   void _tapNodeLink(Node node) {
